fix: Try to cast index and columns of response to int in data_from_df
Sometimes apparently well formatted data arrays were rejected
because of a 'Data are not in generalized trapezoid form.' error.
This could be caused if for example the age index ranges from 1
to 10 but was cast as a string. Then the ordering was incorrect
when pivotting the array, to for example age-cohort space; the
row order would start with age 1, then skip to 10, then to 2, 3
etc. This could cause confusion for '_set_trapezoid_information'
which may then see missing values in positions not admissible
for generalized trapezoids. Trying to cast index and columns to
integers reduces the risk of this issue occuring.